z_9___fastapi/fastapi/requirements.txt
fastapi==0.110.*
uvicorn[standard]==0.29.*
sqlalchemy==2.0.*
pymysql==1.1.*
aiomysql==0.2.*
bcrypt==4.1.*
tensorflow==2.16.*
pillow==10.3.*
python-multipart==0.0.9